Split remaining 8 route domains out of server.py into flux/api mixins
server.py: 5204 -> 1035 lines. Each domain (system, workflows, workers,
admin, schedules, executions, services, rbac) is now a *RoutesMixin module
under flux/api/ following the pattern established by auth_routes.py; handler
bodies moved verbatim, within-domain route order preserved (e.g.
/services/{name}/mcp/tools stays ahead of /services/{name}/{workflow}).

Compatibility kept: schemas + WorkerRegistry re-exported from flux.server so
existing imports and mock.patch targets keep working.

Verified: registered route set byte-identical to the pre-refactor snapshot
(94 routes), full unit suite 2263 passed (same as baseline), pre-commit incl.
mypy green. Version bumped to 0.39.0.
